Moved pilot version banner to auxiliary module
